pyproject.toml
[project]
name = "invoice-data-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.24.0",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "streamlit>=1.41.0",
]
